requirements.txt
Flask==2.0.3
google-cloud-storage==2.3.0
google-cloud-vision==3.0.2
google-cloud-firestore==2.7.2